+++ mlacs/mlas.py
@@ -1,6 +1,9 @@
 """
-// (c) 2021 Aloïs Castellano
-// This code is licensed under MIT license (see LICENSE.txt for details)
+// Copyright (C) 2022-2024 MLACS group (AC, RB)
+// This file is distributed under the terms of the
+// GNU General Public License, see LICENSE.md
+// or http://www.gnu.org/copyleft/gpl.txt .
+// For the initials of contributors, see CONTRIBUTORS.md
 """
 import os
 from concurrent.futures import ThreadPoolExecutor
@@ -91,7 +94,6 @@ class Mlas(Manager):
                  confs_init=None,
                  std_init=0.05,
                  keep_tmp_mlip=True,
-                 ntrymax=0,
                  workdir=''):
 
         Manager.__init__(self, workdir=workdir)
@@ -586,8 +588,8 @@ class Mlas(Manager):
         """
         train_traj, prev_traj = self.read_traj()
 
-        #for i in range(self._nmax):
-        #    self.state[i].subsubdir.mkdir(exist_ok=True, parents=True)
+        # for i in range(self._nmax):
+        #     self.state[i].subsubdir.mkdir(exist_ok=True, parents=True)
 
         # Add the Configuration without a MLIP generating them
         if train_traj is not None: