+++ mlacs/mlip/__init__.py
@@ -1,7 +1,11 @@
 """
-// (c) 2021 Aloïs Castellano
-// This code is licensed under MIT license (see LICENSE.txt for details)
+// Copyright (C) 2022-2024 MLACS group (AC)
+// This file is distributed under the terms of the
+// GNU General Public License, see LICENSE.md
+// or http://www.gnu.org/copyleft/gpl.txt .
+// For the initials of contributors, see CONTRIBUTORS.md
 """
+
 from .mlip_manager import MlipManager
 from .descriptor import Descriptor, SumDescriptor
 from .mliap_descriptor import MliapDescriptor
@@ -11,7 +15,13 @@ from .linear_potential import LinearPotential
 from .delta_learning import DeltaLearningPotential
 from .spin_potential import SpinLatticePotential
 from .mbar_manager import MbarManager
-from .weighting_policy import UniformWeight, IncreasingWeight
+from .ace_descriptor import AceDescriptor
+from .tensorpotential import TensorpotPotential
+from .weights import (UniformWeight,
+                      IncreasingWeight,
+                      EnergyBasedWeight,
+                      FixedWeight)
+
 __all__ = ['MlipManager',
            'Descriptor',
            'SumDescriptor',
@@ -23,4 +33,8 @@ __all__ = ['MlipManager',
            'SpinLatticePotential',
            'MbarManager',
            'UniformWeight',
+           'EnergyBasedWeight',
+           'FixedWeight',
+           'AceDescriptor',
+           'TensorpotPotential',
            'IncreasingWeight',]