+++ mlacs/utilities/io_lammps.py
@@ -1,3 +1,11 @@
+"""
+// Copyright (C) 2022-2024 MLACS group (AC, RB)
+// This file is distributed under the terms of the
+// GNU General Public License, see LICENSE.md
+// or http://www.gnu.org/copyleft/gpl.txt .
+// For the initials of contributors, see CONTRIBUTORS.md
+"""
+
 from pathlib import Path
 
 import numpy as np
@@ -715,6 +723,7 @@ def reconstruct_mlmd_trajectory(trajfile, logfile):
         traj.append(newat)
     return traj
 
+
 # ========================================================================== #
 def get_msd_input(self, msdfile):
     """
@@ -725,8 +734,6 @@ def get_msd_input(self, msdfile):
     for iel, el in enumerate(self.elem):
         block("compute", f"compute c{10+iel} {el} msd com yes")
         block("variable", f"variable msd{el} equal c_c{10+iel}[4]")
-        block("msd el", f"fix f{iel+3} {el} print 1 " + \
-                  f"\"${{msd{el}}}\" screen no append msd{el}.dat")
+        block("msd el", f"fix f{iel+3} {el} print 1 " +
+              f"\"${{msd{el}}}\" screen no append msd{el}.dat")
     return block
-
-