+++ test/RunMLACS.py
@@ -1,3 +1,10 @@
+"""
+// Copyright (C) 2022-2024 MLACS group (AC)
+// This file is distributed under the terms of the
+// GNU General Public License, see LICENSE.md
+// or http://www.gnu.org/copyleft/gpl.txt .
+// For the initials of contributors, see CONTRIBUTORS.md
+"""
 from ase.io import read
 from ase.calculators.lammpsrun import LAMMPS